requirements.txt
Flask==3.0.3
PyYAML==6.0.2